requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
yt-dlp>=2023.10.13
beautifulsoup4>=4.12.2
openai>=1.3.0
pydantic>=2.4.2
sqlalchemy>=2.0.23
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
nltk>=3.8.1
click>=8.1.7 